endec: Decoder: Improve decoder
Due to the fact that some types are decoded differently depending on the
packet and other context information, it was previously created two
different traits, Decoder and DecoderWithContext. While that solution
solved the issue at hand at the moment, it would be inconvenient for
feature use where constraining generics to the Decoder trait would not
be possible. To overcome this limitation, the DecoderWithContext trait
was removed and now only the Decoder trait is available. This was
possible thanks to associated types.

This is still probably not the best solution as most of the times the
context is not required and we're now forced to pass None as argument in
those calls, making it very unecessarily verbose. IMHO still better than
having different traits.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>