control_packets: Improvements
Several improvements to the handling of control packets in order to be a
bit less verbose and drop unnecessary code. Decoder now doesn't require
a context as this was in place only to handle the ConnectPacket
properly and adding a lot of `None`s to decode calls. The crate Result
type is now also used with full qualified name to avoid confusion with
other Result types.

While here, make some stuct fields public and clean up a bit.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>